+++ flutter_sound/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 7.5.0+2
+## 7.5.1+1
 
 - ## The [CHANGELOG file is here](https://www.canardoux.xyz/tau_sound/doc/pages/flutter-sound/api/topics/changelog.html)